--- a/base58_test.go
+++ b/base58_test.go
@@ -83,3 +83,23 @@ func TestDecode(t *testing.T) {
 		}
 	}
 }
+
+func BenchmarkEncode(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		for _, testcase := range testcases {
+			for _, pair := range testcase.testpairs {
+				testcase.encoding.Encode([]byte(pair.decoded))
+			}
+		}
+	}
+}
+
+func BenchmarkDecode(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		for _, testcase := range testcases {
+			for _, pair := range testcase.testpairs {
+				testcase.encoding.Decode([]byte(pair.encoded))
+			}
+		}
+	}
+}