* Added simple benchmarks
* Small optimization for decode/introduce decodeMap
* Optimization for encode/reduce memory allocation

---

Before

```
$ go test -bench .
BenchmarkEncode-8          50000             31670 ns/op
BenchmarkDecode-8          50000             34311 ns/op
PASS
```

After 	1a1d714

```
$ go test -bench .
BenchmarkEncode-8          50000             31852 ns/op
BenchmarkDecode-8          50000             31918 ns/op
PASS
```

After bb0d56b

```
$ go test -bench .
BenchmarkEncode-8         100000             22820 ns/op
BenchmarkDecode-8          50000             31596 ns/op
PASS
```